Moved all Assoc* types to assoc.rs.
